+++ pages/index.vue
@@ -12,7 +12,7 @@
             class="relative px-4 py-2.5 leading-none flex items-center divide-x divide-shark-50 font-semibold text-shark-50"
           >
             <span class="pr-4 text-sm">
-              ✨ Released v1.3.0 on Jan 15 2023 integrates with Raycast for
+              ✨ Released v1.3.0, Supporting integrates with Raycast for
               effective tab searching.
             </span>
             <span